Simplify/improve audio model generation UX
Instead of always showing edit params, with a "Regenerate" button that grays out when nothing has changed,
we jump into the create/edit flow with a "Create/Edit modal model" button.
When that button is pressed, we add a `ModalModelCreateInfo` component.
This opens the modal-model create pane, which can be cancelled to delete the component.
Now, instead of restoring to original params, we just delete the `ModalModelCreateInfo` component.