Match Blender visual behavior when in Edit mode with multiple instances of the same mesh selected:
When multiple instances of the same mesh are selected, only render one of each mesh's instances with Edit visuals.
The rest of the instances are rendered normally (just like non-selected meshes), but now also with the selection silhouette.
This "primary" edit instance is: Active instance, or first selected instance if none active.